WIP Dual Role Port
Things did not go as planned. I attempted to add new features to
the application but with the trace and GUI interface the flash
usage was spent, so I needed to remove things. Unfortunately after
removing trace and GUI I was hard-falting.

I spent time to figure out how to get the LCD working because the
LCD could tell me the information that I wasn't getting from the
GUI so that now works as well as the Joystick.

I also added an external circuit on a breadboard, the poor STM32 GO
no longer looks normal. The breadboard has an enable circuit attached
to a lab bench supply that is currently set up to supply 5 volts.

Now I can start implementing Dual Port Mode.

There are two new signals. An ADC that measures the voltage on the
'source' side of the voltage. As well as an enable signal that can
channel power through the enable circuit. I am concerned, if I make
a mistake in the code I can introduce a voltage to a USB device
that is also attempting to source a voltage. It would likely cause
damage. So I'm proceeding very carefully now.

The next step is to introduce dual port mode within the cube
 and generate multiple source profiles. I'll eventually need to make
some policies on when the devices will supply power and when it will
sink. I'm thinking it will be something like. The sink only devices
are no brainers but if I see a dual role port, I'll need to determine
if I should source power, I can leave this up to the user in some app